generalize the integration test kickoff
1. Renamed integration-test to scheduled-test.
2. Parameterize how the kickoffs are organized.
